Massive Update
Reworked the parsing system so it works much better now. A more extensible type system is in the works. It's actually just how data is stored per value which needs to be changed (to something like a void*) in order to allow a more broader coverage - right now we only allow classes which can be represented as bytes in one way or another. The actual system for custom types is in place. Just specialize the StaticConverter<T> struct and implement the four functions and you should be good. Or wait for the next update where docs for the system will also be released. Finally we implemented a Composer alongside the parser, which composes the HRTDS data into a string which you can store and then reparse again (deserializing).